// src/modules/users/models/IAddressView.ts
export interface IAddressView {
    _id?: string;
    name?: string;
    email?: string;
    mobile: string;
    flat: string;
    landmark: string;
    street: string;
    city: string;
    state: string;
    country: string;
    pinCode: string;
    userObj?: string;
    createdAt?: string;
    updatedAt?: string;
    __v?: number;
}